+++ src/map_events.h
@@ -91,7 +91,7 @@ struct Event {
     unsigned char kind;
     int32_t target;
     /** Button lifespan, decreased over time. When reaches 0, the button disappears. */
-    unsigned long lifespan_turns;
+    uint32_t lifespan_turns;
 };
 
 struct Bookmark {
